+++ stubs/src/main/java/org/bukkit/enchantments/Enchantment.java
@@ -1,0 +1,7 @@
+package org.bukkit.enchantments;
+public class Enchantment {
+ public static final Enchantment FLAME = null;
+ public static final Enchantment INFINITY = null;
+ public static final Enchantment UNBREAKING = null;
+ public static final Enchantment POWER = null;
+}